+++ packages/flutter_riverpod/lib/src/core.dart
@@ -4,3 +4,4 @@ import 'package:meta/meta.dart';
 import 'internals.dart';
 
 part 'core/consumer.dart';
+part 'core/widget_ref.dart';
